pyproject.toml
[project]
name = "adsp-project"
version = "0.1.0"
description = "ADSP Project with FastAPI, Celery, and Postgres"
authors = [{name = "Your Name", email = "your@email.com"}]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.11.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "celery>=5.3.0",
    "redis>=4.6.0",
    "requests>=2.31.0",
    "prometheus-client>=0.17.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    "pydantic-settings>=2.0.0",
    "tenacity>=9.1.2",
]

[project.optional-dependencies]
dev = [
    "invoke>=2.2.0",
    "pytest>=7.4.0",
    "pytest-mock>=3.10.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-requests>=2.0.0",
    "pytest-cov>=4.1.0",
    "watchfiles>=0.21.0"
]

[tool.pytest.ini_options]
addopts = "--cov=app --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    "app/db/base.py",
    "app/main.py",
]

[tool.coverage.report]
fail_under = 90
show_missing = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
